# HEX-LoadBench/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HEX-LoadBench API
  description: |
    Load-Testing & API Stress Test Platform
    
    ## WARNING
    THIS TOOL IS FOR AUTHORIZED PERFORMANCE TESTING ONLY.
    
    Users are responsible for ensuring they have proper authorization before testing any systems.
    
    ## Authentication
    This API uses JWT tokens and API keys for authentication.
    
    - JWT tokens are used for user authentication (session-based)
    - API keys are used for service-to-service communication
    - All endpoints require authentication except `/health` and `/docs`
    
  version: 1.0.0
  contact:
    name: HEX-LoadBench Support
    email: support@hexloadbench.com
    url: https://hexloadbench.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.hexloadbench.com
    description: Production server
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Returns information about the currently authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/overview:
    get:
      tags:
        - Administration
      summary: Get system overview
      description: Returns system statistics and overview (admin only)
      operationId: getSystemOverview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System overview retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemOverview'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/emergency-stop:
    post:
      tags:
        - Administration
      summary: Emergency stop all tests
      description: Stop all running tests immediately (admin only)
      operationId: emergencyStopTests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All tests stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All running tests stopped"
                  stopped_tests:
                    type: integer
                    example: 5
                  stopped_executions:
                    type: integer
                    example: 3
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: number
          format: float
          description: Unix timestamp
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password
          example: "password123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [bearer]
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/UserResponse'
    UserRole:
      type: string
      enum:
        - admin
        - operator
        - viewer
      description: User role in the system
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        email:
          type: string
          format: email
          description: User email
        username:
          type: string
          description: Unique username
        full_name:
          type: string
          description: User's full name
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          description: Whether the user is active
        is_verified:
          type: boolean
          description: Whether the user email is verified
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        last_login:
          type: string
          format: date-time
          description: Last login timestamp
        organization_id:
          type: integer
          description: Organization ID
...
    SystemOverview:
      type: object
      properties:
        users:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            admin:
              type: integer
            inactive:
              type: integer
        tests:
          type: object
          properties:
            total:
              type: integer
            running:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
        system:
          type: object
          properties:
            emergency_kill_switch:
              type: boolean
            database_healthy:
              type: boolean
            version:
              type: string
        recent_activity:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              action:
                type: string
              user_email:
                type: string
              resource_type:
                type: string
              resource_id:
                type: string
...